add new timer by using std::priority_queue fix error VariantData type add cpp version macros
Former-commit-id: 8910fab95e7a577b64e700c6b97f698e6b6e140a
Former-commit-id: d4c78271c76dfcd3d91f2c20744bcbf56016da6a

+++ Frame/SDK/KernelPlugin/AFCElementModule.cpp
@@ -189,7 +189,7 @@ bool AFCElementModule::Load(rapidxml::xml_node<>* attrNode, ARK_SHARE_PTR<AFICla
                 {
                     ARK_ASSERT(0, pTmpNode->name.c_str(), __FILE__, __FUNCTION__);
                 }
-                pTmpNode->value.SetDouble(ARK_LEXICAL_CAST<float>(pstrConfigValue));
+                pTmpNode->value.SetFloat(ARK_LEXICAL_CAST<float>(pstrConfigValue));
             }
             break;
         case DT_DOUBLE:
